Update eyer.sh
more improvements at handling multi files deletion and modification

+++ eyer.sh
@@ -10,7 +10,6 @@ BLUE='\033[0;34m'
 NC='\033[0m'
 
 banner=$(cat << "EOF"
-
    ▄████████ ▄██   ▄      ▄████████    ▄████████    ▄████████ 
   ███    ███ ███   ██▄   ███    ███   ███    ███   ███    ███ 
   ███    █▀  ███▄▄▄███   ███    █▀    ███    █▀    ███    ███ 
@@ -56,25 +55,51 @@ fi
 monitor_file_changes() {
     local file="$1"
     local last_modified=$(date -r "$file" +%s)
+    local deleted=false
     
     while true; do
-        local current_modified=$(date -r "$file" +%s)
-        
-        if [[ "$current_modified" != "$last_modified" ]]; then
-            local current_time=$(date +"%Y-%m-%d %H:%M:%S")
-            local user=$(whoami)
+        if [ ! -f "$file" ]; then
+            if [ "$deleted" = false ]; then
+                local current_time=$(date +"%Y-%m-%d %H:%M:%S")
+                local user=$(whoami)
+                
+                echo -e "${YELLOW}[$current_time] File $file has been deleted by $user!${NC}"
+                echo "[$current_time] $user deleted $file" >> "$LOG_FILE"
+                echo "- File $file has been deleted! at [$current_time] " | notify -id $NOTIFICATION_ID > /dev/null 2>&1
+                
+                deleted=true
+            fi
+        else
+            deleted=false
+            local current_modified=$(date -r "$file" +%s)
             
-            echo -e "${GREEN}[$current_time] File $file has been modified by $user!${NC}"
-            echo "[$current_time] $user made changes in $file" >> "$LOG_FILE"
-            echo "- File $file has been modified! at [$current_time] " | notify -id $NOTIFICATION_ID > /dev/null 2>&1
-            
-            last_modified=$current_modified
+            if [[ "$current_modified" != "$last_modified" ]]; then
+                local current_time=$(date +"%Y-%m-%d %H:%M:%S")
+                local user=$(whoami)
+                
+                echo -e "${GREEN}[$current_time] File $file has been modified by $user!${NC}"
+                echo "[$current_time] $user made changes in $file" >> "$LOG_FILE"
+                echo "- File $file has been modified! at [$current_time] " | notify -id $NOTIFICATION_ID > /dev/null 2>&1
+                
+                last_modified=$current_modified
+            fi
         fi
 
         sleep "$CHECK_INTERVAL"
     done
 }
 
+# Function to handle exit signals
+exit_handler() {
+    pkill -P $$ # Terminate child processes
+    echo -e "${RED}Monitoring terminated.${NC}"
+    exit
+}
+
+# Register exit signals handler
+trap exit_handler EXIT SIGINT SIGTERM
+
+# Start monitoring for each file in the background
 for file in "${FILES_TO_MONITOR[@]}"; do
     if [ ! -f "$file" ]; then
         echo -e "${RED}File '$file' does not exist.${NC}"
